Drawing Debug Lines and Shapes
MDraw can now draw lines, shapes, boxes, n-gons, and circles

+++ Nova/Util/Time.cs
@@ -9,14 +9,14 @@ namespace Nova {
 	public static class Time {
 
 		public static float DeltaTime { get; private set; }
-		public static double TotalTime { get; private set; }
+		public static float TotalTime { get; private set; }
 
 		public static float DeltaDrawTime { get; private set; }
 
 
 		internal static void Update(GameTime time) {
 			DeltaTime = (float)time.ElapsedGameTime.TotalSeconds;
-			TotalTime = time.TotalGameTime.TotalSeconds;
+			TotalTime = (float)time.TotalGameTime.TotalSeconds;
 		}
 
 		internal static void UpdateDraw(GameTime time) {
